xilpm: zynqmp: client: Fix MISRA C-2012 10.1 rule violation
Fix below MISRA C-2012 rule violation:
- Rule 10.1: Operands shall not be of an inappropriate essential type.


Acked-for-series: Izhar Ameer Shaikh <izhar.ameer.shaikh@xilinx.com>

+++ lib/sw_services/xilpm/src/zynqmp/client/common/pm_clock.c
@@ -1036,7 +1036,7 @@ XStatus XPm_GetSelectByClockParent(const enum XPmClock clockId,
 
 	if (NULL == clk->mux->inputs) {
 		/* MIO mux */
-		u32 mioSel = parentId - PM_CLOCK_EXT_MIO0;
+		u32 mioSel = (u32)parentId - (u32)PM_CLOCK_EXT_MIO0;
 		if (mioSel <= 0x4DU) {
 			*select = mioSel;
 			status = (XStatus)XST_SUCCESS;
